+++ tools/src/commands/SetupReactNativeNightly.ts
@@ -1,6 +1,6 @@
 import { Command } from '@expo/commander';
 import JsonFile from '@expo/json-file';
-import fs from 'fs-extra';
+import glob from 'glob-promise';
 import path from 'path';
 
 import { EXPO_DIR } from '../Constants';
@@ -33,10 +33,15 @@ async function main() {
 
   await updateReactNativePackageAsync();
 
-  await patchReanimatedAsync();
+  await patchReanimatedAsync(nightlyVersion);
+
+  await removeKotlinAndroidExtensionAsync();
+
+  logger.info('Setting up Expo modules files');
+  await updateExpoModulesAsync();
 
   logger.info('Setting up project files for bare-expo.');
-  await updateBareExpoAsync();
+  await updateBareExpoAsync(nightlyVersion);
 }
 
 async function addPinnedPackagesAsync(packages: Record<string, string>) {
@@ -51,84 +56,215 @@ async function addPinnedPackagesAsync(packages: Record<string, string>) {
 }
 
 async function updateReactNativePackageAsync() {
-  const root = path.join(EXPO_DIR, 'node_modules', 'react-native');
-
-  // Third party libraries used to use react-native minor version, update the version 9999.9999.9999 as the latest version
-  await transformFileAsync(path.join(root, 'package.json'), [
-    {
-      find: '"version": "0.0.0-',
-      replaceWith: '"version": "9999.9999.9999-',
-    },
-  ]);
-  await transformFileAsync(path.join(root, 'ReactAndroid', 'gradle.properties'), [
-    {
-      find: 'VERSION_NAME=0.0.0-',
-      replaceWith: 'VERSION_NAME=9999.9999.9999-',
-    },
-  ]);
-
-  // Build hermes source from the main branch
-  await fs.writeFile(path.join(root, 'sdks', '.hermesversion'), 'main');
-  await transformFileAsync(path.join(root, 'sdks', 'hermes-engine', 'hermes-engine.podspec'), [
-    {
-      // Because we changed the version in package.json, the `isNightly` check in hermes-engine.podspec is broken
-      find: "isNightly = version.start_with?('0.0.0-')",
-      replaceWith: 'isNightly = true',
-    },
-  ]);
-
-  // Remove unused hermes build artifacts to reduce build time
-  await transformFileAsync(path.join(root, 'sdks', 'hermes-engine', 'hermes-engine.podspec'), [
-    {
-      find: './utils/build-mac-framework.sh',
-      replaceWith: '',
-    },
-  ]);
+  // Workaround for react-native-gradle-plugin
+  const gradlePluginRoot = path.join(EXPO_DIR, 'node_modules', 'react-native-gradle-plugin');
   await transformFileAsync(
-    path.join(root, 'sdks', 'hermes-engine', 'utils', 'build-ios-framework.sh'),
+    path.join(gradlePluginRoot, 'src/main/kotlin/com/facebook/react/ReactExtension.kt'),
     [
       {
-        find: 'build_apple_framework "iphoneos" "arm64" "$ios_deployment_target"',
+        find: 'internal val reactNativeDir:',
+        replaceWith: 'val reactNativeDir:',
+      },
+    ]
+  );
+
+  const reactNativeRoot = path.join(EXPO_DIR, 'node_modules', 'react-native');
+  // Workaround duplicated libc++_shared.so from linked fbjni
+  await transformFileAsync(path.join(reactNativeRoot, 'ReactAndroid', 'build.gradle'), [
+    {
+      find: /^(\s*packagingOptions \{)$/gm,
+      replaceWith: '$1\n        pickFirst("**/libc++_shared.so")',
+    },
+  ]);
+}
+
+async function patchReanimatedAsync(nightlyVersion: string) {
+  const root = path.join(EXPO_DIR, 'node_modules', 'react-native-reanimated');
+
+  await transformFileAsync(path.join(root, 'scripts', 'reanimated_utils.rb'), [
+    // Add REACT_NATIVE_OVERRIDE_VERSION support
+    {
+      find: `result[:react_native_version] = react_native_json['version']`,
+      replaceWith: `result[:react_native_version] = ENV["REACT_NATIVE_OVERRIDE_VERSION"] ? ENV["REACT_NATIVE_OVERRIDE_VERSION"] : react_native_json['version']`,
+    },
+    {
+      find: `result[:react_native_minor_version] = react_native_json['version'].split('.')[1].to_i`,
+      replaceWith: `result[:react_native_minor_version] = result[:react_native_version].split('.')[1].to_i`,
+    },
+  ]);
+  await transformFileAsync(path.join(root, 'android', 'build.gradle'), [
+    // Add REACT_NATIVE_OVERRIDE_VERSION support
+    {
+      find: `def REACT_NATIVE_VERSION = reactProperties.getProperty("VERSION_NAME")`,
+      replaceWith: `def REACT_NATIVE_VERSION = System.getenv("REACT_NATIVE_OVERRIDE_VERSION") ?: reactProperties.getProperty("VERSION_NAME")`,
+    },
+    // Workaround $minor is undefined
+    {
+      find: /\$minor/g,
+      replaceWith: '$rnMinorVersion',
+    },
+    // BUILD_FROM_SOURCE
+    {
+      find: /^(boolean BUILD_FROM_SOURCE)\s*=.*/gm,
+      replaceWith: '$1 = true',
+    },
+    // duplicated class from jni, because ReactAndroid now uses fbjni rather than fbjni-java-only
+    {
+      find: 'implementation "com.facebook.fbjni:fbjni-java-only:',
+      replaceWith: 'compileOnly "com.facebook.fbjni:fbjni:',
+    },
+    {
+      // no-op tasks
+      find: /\b(task (prepareHermes).*\{)$/gm,
+      replaceWith: `$1\n    return`,
+    },
+    {
+      // download nightly react-native aar
+      find: /^(task unpackReactNativeAAR \{[\s\S]*?^\})/gm,
+      replaceWith: `
+def reactNativeIsNightly = reactProperties.getProperty("VERSION_NAME").startsWith("0.0.0-")
+
+def downloadReactNativeNightlyAAR = { buildType, version, downloadFile ->
+  def classifier = buildType == 'Debug' ? 'debug' : 'release'
+  download.run {
+    src("https://oss.sonatype.org/service/local/artifact/maven/redirect?r=snapshots&g=com.facebook.react&a=react-native&c=\${classifier}&e=aar&v=\${version}-SNAPSHOT")
+    onlyIfNewer(true)
+    overwrite(false)
+    dest(downloadFile)
+  }
+}
+
+task unpackReactNativeAAR {
+  def buildType = resolveBuildType()
+  def rnAAR
+  if (reactNativeIsNightly) {
+    def downloadFile = file("\${downloadsDir}/react-native-nightly.aar")
+    downloadReactNativeNightlyAAR(buildType, reactProperties.getProperty("VERSION_NAME"), downloadFile)
+    rnAAR = downloadFile
+  } else {
+    def rnAarMatcher = "**/react-native/**/*\${buildType}.aar"
+    if (REACT_NATIVE_MINOR_VERSION < 69) {
+      rnAarMatcher = "**/**/*.aar"
+    }
+    rnAAR = fileTree("$reactNativeRootDir/android").matching({ it.include rnAarMatcher }).singleFile
+  }
+  def file = rnAAR.absoluteFile
+  def packageName = file.name.tokenize('-')[0]
+  copy {
+    from zipTree(file)
+    into "$reactNativeRootDir/ReactAndroid/src/main/jni/first-party/$packageName/"
+    include "jni/**/*.so"
+  }
+}
+      `,
+    },
+    {
+      // add prefab support, setup task dependencies and hermes-engine dependencies
+      transform: (text: string) =>
+        text +
+        '\n\n' +
+        `android {\n` +
+        `  buildFeatures {\n` +
+        `    prefab true\n` +
+        `  }\n` +
+        `}\n` +
+        `\n` +
+        `dependencies {\n` +
+        `  compileOnly "com.facebook.react:hermes-engine:${nightlyVersion}-SNAPSHOT"` +
+        `}\n`,
+    },
+  ]);
+
+  await transformFileAsync(path.join(root, 'android', 'CMakeLists.txt'), [
+    {
+      // Remove this after reanimated support react-native 0.71
+      find: /(\s*"\$\{REACT_NATIVE_DIR\}\/ReactAndroid\/src\/main\/jni")/g,
+      replaceWith: '$1\n        "${REACT_NATIVE_DIR}/ReactAndroid/src/main/jni/react/turbomodule"',
+    },
+    {
+      // find hermes from prefab
+      find: /(string\(APPEND CMAKE_CXX_FLAGS " -DJS_RUNTIME_HERMES=1"\))/g,
+      replaceWith: `find_package(hermes-engine REQUIRED CONFIG)\n    $1`,
+    },
+    {
+      // find hermes from prefab
+      find: /"\$\{BUILD_DIR\}\/.+\/libhermes\.so"/g,
+      replaceWith: `hermes-engine::libhermes`,
+    },
+  ]);
+
+  // Workaround for UIImplementationProvider breaking change, that would break reanimated layout animation somehow
+  await transformFileAsync(
+    path.join(
+      root,
+      'android/src/main/java/com/swmansion/reanimated/layoutReanimation/ReanimatedUIManager.java'
+    ),
+    [
+      {
+        find: /^class ReaUiImplementationProvider extends UIImplementationProvider \{[\s\S]*?^\}/gm,
         replaceWith: '',
       },
       {
-        find: 'build_apple_framework "catalyst" "x86_64;arm64" "$ios_deployment_target"',
+        find: `new ReaUiImplementationProvider(),`,
         replaceWith: '',
-      },
-      {
-        find: 'create_universal_framework "iphoneos" "iphonesimulator" "catalyst"',
-        replaceWith: 'create_universal_framework "iphonesimulator"',
       },
     ]
   );
 }
 
-async function patchReanimatedAsync() {
-  // Workaround for reanimated doesn't support the hermes where building from source
-  const root = path.join(EXPO_DIR, 'node_modules', 'react-native-reanimated');
-  await transformFileAsync(path.join(root, 'android', 'build.gradle'), [
-    {
-      find: /\bdef hermesAAR = file\(.+\)/g,
-      replaceWith:
-        'def hermesAAR = file("$reactNative/ReactAndroid/hermes-engine/build/outputs/aar/hermes-engine-debug.aar")',
-    },
-  ]);
+/**
+ * Remove deprecated kotlin-android-extensions
+ * TODO: remove this after detox updated
+ */
+async function removeKotlinAndroidExtensionAsync() {
+  const gradleFiles = ['node_modules/detox/android/detox/build.gradle'];
 
-  // Remove this after reanimated support react-native 0.71
-  await transformFileAsync(path.join(root, 'android', 'CMakeLists.txt'), [
-    {
-      find: /(\s*"\$\{NODE_MODULES_DIR\}\/react-native\/ReactAndroid\/src\/main\/jni")/g,
-      replaceWith:
-        '$1\n        "${NODE_MODULES_DIR}/react-native/ReactAndroid/src/main/jni/react/turbomodule"',
-    },
-  ]);
+  await Promise.all(
+    gradleFiles.map((file) =>
+      transformFileAsync(file, [
+        {
+          find: /apply plugin: ['"]kotlin-android-extensions['"]/g,
+          replaceWith: '',
+        },
+      ])
+    )
+  );
 }
 
-async function updateBareExpoAsync() {
-  const gradlePropsFile = path.join(EXPO_DIR, 'apps', 'bare-expo', 'android', 'gradle.properties');
-  let content = await fs.readFile(gradlePropsFile, 'utf8');
-  if (!content.match('reactNativeNightly=true')) {
-    content += `\nreactNativeNightly=true\n`;
-    await fs.writeFile(gradlePropsFile, content);
-  }
+async function updateExpoModulesAsync() {
+  const gradleFiles = await glob('packages/**/build.gradle', { cwd: EXPO_DIR });
+  await Promise.all(
+    gradleFiles.map((file) =>
+      transformFileAsync(file, [
+        {
+          find: /\b(com.facebook.fbjni:fbjni):0\.2\.2/g,
+          replaceWith: '$1:0.3.0',
+        },
+      ])
+    )
+  );
+}
+
+async function updateBareExpoAsync(nightlyVersion: string) {
+  const root = path.join(EXPO_DIR, 'apps', 'bare-expo');
+  await transformFileAsync(path.join(root, 'android', 'build.gradle'), [
+    {
+      find: 'resolutionStrategy.force "com.facebook.react:react-native:${reactNativeVersion}"',
+      replaceWith: `resolutionStrategy.force "com.facebook.react:react-native:${nightlyVersion}-SNAPSHOT"`,
+    },
+  ]);
+
+  await transformFileAsync(path.join(root, 'ios', 'BareExpo', 'AppDelegate.mm'), [
+    {
+      // Remove this when we upgrade bare-expo to 0.71
+      find: `  RCTAppSetupPrepareApp(application);`,
+      replaceWith: `
+#if RCT_NEW_ARCH_ENABLED
+  RCTAppSetupPrepareApp(application, YES);
+#else
+  RCTAppSetupPrepareApp(application, NO);
+#endif
+`,
+    },
+  ]);
 }